Refactored code, finished menu
